+++ web/public/index.html
@@ -10,10 +10,7 @@
             content="minimum-scale=1, initial-scale=1, width=device-width"
         />
 
-        <meta
-            name="description"
-            content="Web site created using create-react-app"
-        />
+        <meta name="description" content="Servus" />
         <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
         <!--
       manifest.json provides metadata used when your web app is installed on a
