+++ web/src/pages/Home.js
@@ -1,7 +1,9 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import Map from '../components/map/Map';
 import { makeStyles } from '@material-ui/core/styles';
 import ServiceList from '../components/ServiceList';
+import { useCookies } from 'react-cookie';
+import { useHistory } from 'react-router-dom';
 
 const useStyles = makeStyles(theme => ({
     container: {
@@ -19,8 +21,12 @@ const useStyles = makeStyles(theme => ({
 
 const Home = () => {
     const classes = useStyles();
+    const [cookies] = useCookies();
+    const history = useHistory();
     const [selected_service, setService] = useState();
-
+    useEffect(() => {
+        if (!cookies.username) history.push('/login');
+    }, []);
     return (
         <div className={classes.container}>
             <ServiceList
